uprobes: split into enable and patch fragments
To allow systemtap and the standard kernel to unconditionally include
uprobes we split it into the configuration options to enable it (for
systemtap's use) and into the patches (for the standard kernel to
use). This allows both the feature fragment and the standard kernel
to enable systemtap, but not double patch the kernel.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>