Adde "early debug printk" support.
The changes to init/main.c are **TEMPORARY**.  As soon
as we start booting with U-Boot, they won't be needed.

At present, the early printk debug support assumes that the
physical address of the UART is a 32 bit value.  This is not
the case for Axxia.  Until there is a way to do this...

Signed-off-by: John Jacques <john.jacques@lsi.com>