LSI axm55xx: Add multi-cluster support for up to 16 cores
The LSI axm55xx platform can have up to four clusters, each having four
A15 cores (so up to a total of 16 cores). The current setup assumes that
a single GIC will handle the IPI interrupts on a platform, but a single
GIC can only handle up to eight cores. On the axm55xx platform, an
external distributed interrupt IPI block is used in place of the GIC
software generated IPIs. In this commit, we replace the use of the
standard GIC driver with a axm55xx specific one that supports multiple
clusters.

Signed-off-by: David Mercado <david.mercado@windriver.com>
Signed-off-by: Wang Hui <Hui.Wang@windriver.com>